<R11698> Reverting to an earlier version (-r7388) that has been running on Bach for all of these years.  The latest version contains changes that created a "head" of a group for any Summary whose SummaryURL changed.
That never went into production and so is not fully tested.  I am reverting to
the version actually in production in order to avoid putting untested code on
the CBIIT servers, and to ensure that if CDR software is exported from version
control, the untested version will not be exported.